Rename package and folder api/client to api/core
