+++ api/core/execute.go
@@ -1,4 +1,4 @@
-package client
+package core
 
 import (
 	"encoding/json"
