+++ api/core/execute_test.go
@@ -1,4 +1,4 @@
-package client
+package core
 
 import (
 	"context"
