+++ api/core/listen_event.go
@@ -1,4 +1,4 @@
-package client
+package core
 
 import (
 	"encoding/json"
