+++ cmd/service/test.go
@@ -6,7 +6,7 @@ import (
 	"log"
 	"time"
 
-	"github.com/mesg-foundation/core/api/client"
+	"github.com/mesg-foundation/core/api/core"
 
 	"github.com/mesg-foundation/core/cmd/utils"
 	"github.com/mesg-foundation/core/config"
@@ -34,16 +34,16 @@ mesg-cli service test --keep-alive`,
 	DisableAutoGenTag: true,
 }
 
-func startServiceFromCli(cli client.CoreClient, service *service.Service) {
-	_, err := cli.StartService(context.Background(), &client.StartServiceRequest{
+func startServiceFromCli(cli core.CoreClient, service *service.Service) {
+	_, err := cli.StartService(context.Background(), &core.StartServiceRequest{
 		Service: service,
 	})
 	handleError(err)
 	log.Println("Service", service.Name, "started")
 }
 
-func listenEvents(cli client.CoreClient, service *service.Service, filter string) {
-	stream, err := cli.ListenEvent(context.Background(), &client.ListenEventRequest{
+func listenEvents(cli core.CoreClient, service *service.Service, filter string) {
+	stream, err := cli.ListenEvent(context.Background(), &core.ListenEventRequest{
 		Service: service,
 	})
 	handleError(err)
@@ -59,8 +59,8 @@ func listenEvents(cli client.CoreClient, service *service.Service, filter string
 	}
 }
 
-func listenResults(cli client.CoreClient, service *service.Service) {
-	stream, err := cli.ListenResult(context.Background(), &client.ListenResultRequest{
+func listenResults(cli core.CoreClient, service *service.Service) {
+	stream, err := cli.ListenResult(context.Background(), &core.ListenResultRequest{
 		Service: service,
 	})
 	handleError(err)
@@ -74,7 +74,7 @@ func listenResults(cli client.CoreClient, service *service.Service) {
 	}
 }
 
-func executeTask(cli client.CoreClient, service *service.Service, task string, dataPath string) (execution *client.ExecuteTaskReply, err error) {
+func executeTask(cli core.CoreClient, service *service.Service, task string, dataPath string) (execution *core.ExecuteTaskReply, err error) {
 	if task == "" {
 		return
 	}
@@ -84,7 +84,7 @@ func executeTask(cli client.CoreClient, service *service.Service, task string, d
 		handleError(err)
 	}
 
-	execution, err = cli.ExecuteTask(context.Background(), &client.ExecuteTaskRequest{
+	execution, err = cli.ExecuteTask(context.Background(), &core.ExecuteTaskRequest{
 		Service:  service,
 		TaskKey:  task,
 		TaskData: string(data),
@@ -99,7 +99,7 @@ func testHandler(cmd *cobra.Command, args []string) {
 
 	connection, err := grpc.Dial(viper.GetString(config.APIClientTarget), grpc.WithInsecure())
 	handleError(err)
-	cli := client.NewCoreClient(connection)
+	cli := core.NewCoreClient(connection)
 
 	startServiceFromCli(cli, service)
 	defer service.Stop()
